Fix: Formatting for parameter default values (#3946)
This PR fixes #3944. I first recreated the issue with a test, and then
fixed it.
It turned out, formals had a field "DefaultValue" which was never
considered as a child, so its tokens were considered to be part of the
declaration, which messed up indentation.
The fix was to ensure Formals now take the default value also in their
range, and also in their children.