First draft of verification optimization guide (#3934)
Adds a new document describing techniques for optimizing verification to
reduce resource use and variability.

This in some ways overlaps with the sections in the manual about fixing
verification failures, optimizing slow verification, and reducing variability.
I'd ultimately like it to supersede those, once we ensure that we're not
losing material that way. Those DafnyRef sections can be replaced by a
link to this new document.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.