Remove IncludeToken as prep for caching parsing (#3858)
Currently a file that's included is parsed differently. Its tokens are
wrapped in an `IncludeToken`, that includes a reference to the including
file.

To allow reusing parsed files in a cache, how the file is parsed should
be independent of what triggered it to be parsed. To achieve that, this
PR removes `IncludeToken`. To enable still determining whether a
particular token was part of an include or not, tokens now store a `Uri`
that uniquely identifies the document that token came from. The Uris of
the root documents are stored in `DefaultModuleDefinition`, which is
just below `Program`. Any token with a Uri that is not part of the root
set is considered to be included.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>