Specify target type explicitly for type descriptors in Java
Otherwise Java tries to infer it based on the arguments and may get it wrong, since the .class literal can’t be of the exact target type (e.g. List.class is legal, List<T>.class is not)

Similarly, change TypeDescriptor.boxedClass from `Class<T>` to `Class<?>`, since it’s not possible to be an exact match in all cases.