refactor: Improve ClassDecl hierarchy (#4025)
This PR changes the class hierarchy of the classes that previously
rolled up to `ClassDecl`. This both improves the organization of the
code and paves the way for traits to move away from always being
reference types.

The old class hierarchy looked like:

```
   - TopLevelDecl
     - TopLevelDeclWithMembers
       - ValuetypeDecl
       - ClassDecl
         - ArrayClassDecl
         - IteratorDecl
         - DefaultClassDecl
         - TraitDecl
         - ArrowTypeDecl
```

Of these classes, `ArrowTypeDecl` and `DefaultClassDecl` are not really
classes, but were historically placed under `ClassDecl` when `ClassDecl`
was the only thing that contained members. (A further improvement, not
in this PR, would be to rename `DefaultClassDecl` to something like
`ModuleMembers`.) Also, while it previously was okay to treat
`TraitDecl` as a kind of `ClassDecl`, this is about to change in Dafny,
since traits will soon be available for non-reference types, too.

The new class hierarchy looks like:

```
   - TopLevelDecl
     - TopLevelDeclWithMembers
       - ValuetypeDecl
         - ArrowTypeDecl
       - DefaultClassDecl
       - ClassLikeDecl
         - ClassDecl
           - ArrayClassDecl
           - IteratorDecl
         - TraitDecl
```

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>