Creating a dafny doc tool (#3849)
`dafny doc <files-and-folders>`

processes all the .dfy files on the commandline or in the folders
(recursively) on the command line,
creating html 'dafny-doc' pages summarizing and describing the contents
of the files, using the docstrings on the declarations.

Only operates if all files pass parsing and resolution (and all
command-line options relevant to those phases apply here also)

`--output:<folder>` -- the folder in which the output will be placed (a
flat collection of .html files). Default is './docs'. Folder is created
if needed; contents are overwritten, but not deleted.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: davidcok <davidcok@github.com>
Co-authored-by: Mikaël Mayer <MikaelMayer@users.noreply.github.com>