Cache parsing (#4085)
### Changes
- Let the Dafny language server use a cache for parsing. The cache
stores items as long as they were used in the previous compilation. This
means that commenting out an include can cause a large piece of cache to
be pruned, leading to a possibly unexpected slowdown when it's
uncommented again.

### Caveats
I've decided not to add an option to turn off caching, since I don't see
the use-case. Of course caching is a computation vs storage trade-off,
and we could provide many options to configure the caching behavior, but
I think the current caching is cheap enough (at most storing two sets of
parse results at a time) that there is no reason to want to turn it off.

Please carefully review the code that determines the cache key, since a
bug in that would lead to outdated compilation results.

### Testing
- Add a test that checks if parsing is cached, and that the cache prunes
entries when they're not hit.
- Add test for included file has error message for indirect
includes, something that was still missing from
https://github.com/dafny-lang/dafny/pull/4083

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>