Disable ShowSnippets during Test Generation (#4216)
The `--show-snippets` breaks test generation because it relies on
casting of `Boogie.IToken` to `Dafny.IToken`
[here](https://github.com/dafny-lang/dafny/blob/d72dc7a85ae5d662dda78f927463d1e56b4a0696/Source/DafnyCore/DafnyConsolePrinter.cs#L54)
and test generation makes changes to the Boogie AST that may introduce
new tokens that cannot be cast. This PR disables the option, recently
enabled by default, during test generation. An alternative would be to
surround the cast with try-catch, but it seems that outside test
generation purposes it is probably safe to assume that the cast
succeeds. @atomb, @robin-aws would it be possible to merge this before
the next release to make sure test generation stays stable?


By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>