Do not publish an error saying module resolution is not done because … (#4237)
Preparation for https://github.com/dafny-lang/dafny/pull/3902

### Changes

Error messages of the form `ErrorsInRelatedModules.dfy(135,9): Error:
not resolving module 'ClientOfErroneousModule5' because there were
errors in resolving its import 'ModuleWithErrors'`, are no longer
emitted. The user can expect that some resolution errors will block the
resolution of other parts of the program, so this error is not useful.
Moreover, this error can produce too much noise because it can cascade.
Suppose you have 1 module with an error and a large tree of modules that
are importing this one module. Previously all of the modules would
produce an error and only 1 of those would be useful, now only that
useful one is shown.

### Testing
Updated existing tests

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>