Issue 4035 revisited (#4220)
Fixes #4035 
Fixes #4217
Fixes #4224

There were two problems with the previous fix of Issue #4035 (PR #4036,
which was reverted in PR #4192), both related to abstract types. One
problem was that the previous fix used `$Is`, not `$IsBox`, for abstract
types. The other problem was that Dafny has (for a long time, it seems)
generated two Boogie type terms for an abstract type, one of the form
`#$MyAbstractType` and the other of the form `TMyAbstractType()`. This
PR eliminates the former form. As part of this change, the
`#$InternalTypeSynonym` type terms have now also changed to
`TInternalTypeSynonym()`.

In addition, this PR fixes a soundness problem (#4224), where
injectivity axioms had been generated for export-provided types (which
shouldn't happen).

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>