Fix: Modules verified in the correct order to prevent Boogie Crash (#4212)
This PR fixes #4139
I'm not sure of the fix right now, need to see if it works on other
tests. I suspect it will fail and we might need another way to separate
the TopLevelDecls order for verification as well. But let's see.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>