Don't say "error is impossible" anymore
This was a confusing message and it's possible to do better.

This addresses #3954, but I don't think it entirely resolves it. To do
that, I think it'd be best to extend ProofObligationDescription to
include related locations, which will require Boogie changes.

This change helps the situation in the meantime, however.