feat: Compile general traits with type parameters (#4240)
This PR implements support in all 5 compilers for type parameters in
types that extend traits. In particular, the compilation support covers
datatypes and codatatypes that extend traits. This functionality is
available as part of the `/generalTraits:1` beta.

(After this PR, there will be one more PR to complete compilation
support for general traits. It has to do with `newtype`s that extend
traits.)

(Note, to use datatype as the traits they extend, it is necessary to use
not only `/generalTraits:1` but also `/typeSystemRefresh:1`. The latter
will be available in a future PR.)

## More details

Here follows a description of what this PR implements and why.

Consider the following types, which are already supported by Dafny:

``` dafny
class Class<T(0)> { // note the auto-init type parameter
  method M() {
    var t: T; // the compiler auto-initializes "t", since "T" is an auto-init type
    print t;
  }
}

datatype Datatype<T(0)> = Datatype { // note the auto-init type parameter
  method M() {
    var t: T; // the compiler auto-initializes "t", since "T" is an auto-init type
    print t;
  }
}
```

The implementations of methods `Class.M()` and `Datatype.M()` need a
type descriptor for `T`. Before this PR, where this type descriptor
comes from is different from a class and a datatype.

For the class, the type descriptor is stored in each object. Let's call
it "Design A". This design makes the type descriptor easily accessible
from the implementation of `Class.M()` and saves the caller from having
to create such a type descriptor with each call to `M()`. On the
downside, it means that the type descriptor occupies space in each
object.

For the datatype, the design is the opposite. Let's call it "Design B".
Under this design, the type parameter is passed into `Datatype.M()` as a
parameter at run time, which has the pros/cons of Design A's cons/pros.

To support traits and dynamic dispatch, Design B is no longer an option.
Consider the following variation of the types above:

``` dafny
trait Parent {
  method M()
}

class Class<T(0)> extends Parent {
  method M() {
    var t: T;
    print t;
  }
}

datatype Datatype<T(0)> extends Parent = Datatype {
  method M() {
    var t: T;
    print t;
  }
}

method Main() {
  var c: Class<int> := new Class<int>;
  var d: Datatype<bool> := Datatype;
  var p: Parent;
  p := c;
  p.M();
  p := d;
  p.M();
}
```

Here, the calls to `Parent.M()` (which are dynamically dispatched to
`Caller.M()` and `Datatype.M()`, respectively) do not take any
type-descriptor parameters, because there are no type parameters
involved in `Parent.M()`. Hence, the only way for an implementation of
`M()` in a child type that extends `Parent` to get hold of the necessary
type parameters (which are specific to the child type) is for the
child-type value to contain the type descriptors. So, we must choose
Design A for any type (class, datatype, ...) that can extend a trait.

So, this PR changes the run-time representation of (co)datatypes to
include, in the (co)datatype values themselves, the necessary type
descriptors. While there is a cost to this at run time, it is not common
for a (co)datatype to have an auto-init type parameter, so the target
code in most cases will be unchanged. _Except_ for Java, which clumsily
adds type descriptors for _all_ type parameters (not just auto-init type
parameters). For several reasons (for example, see
https://github.com/dafny-lang/dafny/issues/3055), this would be good to
change in the (hopefully near) future.

This change also affects the "downward clone" mechanism used by the C#
target.

## Alternative

Rather than adding type descriptors as fields in the run-time
representation of classes and datatypes, one could include such type
descriptors in the vtable that each such class/datatype already has a
pointer to at run time. However, there are two issues that make such a
change difficult. One issue is that such vtables are controlled by the
target language. Thus, for each required instantiation of the type
descriptors, one would need to declare a separate leaf type in the
target language. The other issue is that the required instantiations may
be difficult to compute statically.

So, while this alternative may be considered in the future (especially
for a native compiler for Dafny), it is out of scope for this PR.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Remy Willems <rwillems@amazon.com>