Prevent a race condition related to ProjectManagerDatabase.OpenDocument (#4303)
### Changes
- Update `DafnyTextDocumentHandler.cs` so it properly awaits
`ProjectManagerDatabase.OpenDocument`, so no new requests can be handled
until the ProjectManager's constructor and `OpenDocument` method have
been called. Without this change, a call to `CreateAndOpenTestDocument`
could deadlock, because the hover request would come in while the
OpenDocument request would still create a new compilation, causing the
hover to wait for a compilation that was never started. It's not clear
why the hover wouldn't return as cancelled.
- Use `LanguageProtocolTestBase` instead of `LanguageServerTestBase`,
which means more code re-use from the library
- Do not Dispose the entire `ServiceCollection` but only the
`ProjectManagerDatabase`, which is safer because we're disposing less
but still enough. Hopefully this leads to fewer exceptions in the test
runs.

### Testing
- Add a test `UpdateDuringARequestWillCancelTheRequest`

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>