Update to Boogie 3.0.0 (#4209)
This adds a couple of Dafny-relevant things:
* A soundness fix for boogie-org/boogie#749. I thought this would resolve #4053, because it resolves a soundness issue discovered during the investigation of that issue, but it doesn't. Apparently there's _another_ soundness issue hiding in there.
* Support for coverage analysis: which program constructs annotated with
an `{:id ...}` attribute contributed to verification? Dafny currently
does not add `{:id ..}` attributes, but we plan to make it automatically
do so (optionally) in the near future. See boogie-org/boogie#730.
* Allow absolute paths to Boogie solver plugins.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.