Make the Boogie representation of the Dafny heap use only Box (#4020)
This PR makes a simple but significant change to the Boogie
representation of the Dafny heap, ensuring that all values stored in it
are of type `Box`. This makes the polymorphism in the Boogie `$Heap`
variable irrelevant, but does not remove it. Removing it will,
unfortunately, require slightly more significant changes to Dafny
itself, more than the prelude. I plan that in a separate PR, but this
one already seems to reduce resource use and resource use variation a
little bit, so I thought it'd be good to merge it now. These changes
seem to work best if done in very small steps.

The `ReadHeap` method now takes an optional type, which will ultimately
be provided at all call sites. Once it is, the `Field` type in the
prelude will no longer need a parameter and the heap can become truly
monomorphic. The correct value to pass in isn't obvious at all call
sites, however, so I'm filling each one in incrementally as I conclude
what the right parameter should be.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.

---------

Co-authored-by: Aleksandr Fedchin <fedchina@amazon.com>