fix: Continue to compilation tests after comparing .verifier.expect (#4356)
This PR fixes and changes the behavior of `%testDafnyForEachCompiler`
tests. The bug that is fixed had caused 16 compiler tests in the Dafny
test suite to not run.

Here is a description of the new behavior, where _italicized_ parts
indicate changes from before:

* Verification is expected to go through with a 0 exit code.
* If there is a `.verifier.expect` file, then the actual output is
expected to contain a blank line followed by a line that starts with
`Dafny program verifier`, and this portion of the output will be removed
before comparing the output. The remaining output is compared with the
contents of the `.verifier.expect` file and is expected to match.
* If there is no `.verifier.expect` file, the verifier output minus the
`Dafny program verifier` line (see previous bullet) is expected to be
empty.
* When running the test for each compiler, if the output contains a line
starting with `Dafny program verifier`, then _everything up to and
include that line is removed_ before comparing the output.

This PR adds add test in `Test/metatests`.

Fixes #4355

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>

---------

Co-authored-by: Robin Salkeld <salkeldr@amazon.com>