Set options in IDE to more closely match CLI (#4374)
This makes IDE and CLI resource counts more likely to match. In
particular, it makes it so that the Boogie file generated when running
the language server with the `--bprint` flag matches the Boogie file
generated when running the CLI with the `/print` Boogie flag, except for
some comments that don't seem to affect RU counts.

This does not, unfortunately, always make the RU counts presented in the
IDE match those given by the CLI. I'm not yet sure where the additional
discrepancy is coming from.

By submitting this pull request, I confirm that my contribution
is made under the terms of the MIT license.